Fix issue with crash on LibreOffice format.
During tests, there was issue regarding ODF2XHTML.xhtml() method, which
assumes, that all lines are stored as strings. Although, in case of
single items, like footnotes, there is a chance, that numbers are stored
as ints. In this patch we assure, that all lines has to be strings. This
bug also affects current version of Calibre (as of cf5baaf449 commit).