Cut off cross importing things.
There was (and perhaps it exists in other modules, need to check) a bad
habit of importing objects out of module A into B just to be able of
import those object from module B. In this case there was import from
module config_base into module config. This commit fix this.